+++ src/main/java/net/frozenblock/wilderwild/mod_compat/BetterEndIntegration.java
@@ -21,8 +21,8 @@ package net.frozenblock.wilderwild.mod_compat;
 import java.util.function.BooleanSupplier;
 import net.frozenblock.lib.integration.api.ModIntegration;
 import static net.frozenblock.lib.sound.api.block_sound_group.BlockSoundGroupOverwrites.addBlock;
+import static net.frozenblock.wilderwild.registry.WWSoundTypes.*;
 import net.frozenblock.wilderwild.config.WWBlockConfig;
-import static net.frozenblock.wilderwild.registry.WWSoundTypes.LEAVES;
 
 public class BetterEndIntegration extends ModIntegration {
 	public BetterEndIntegration() {
@@ -31,12 +31,39 @@ public class BetterEndIntegration extends ModIntegration {
 
 	@Override
 	public void init() {
-		BooleanSupplier condition = () -> WWBlockConfig.get().blockSounds.leafSounds;
-		addBlock(id("pythadendron_leaves"), LEAVES, condition);
-		addBlock(id("lacugrove_leaves"), LEAVES, condition);
-		addBlock(id("dragon_tree_leaves"), LEAVES, condition);
-		addBlock(id("tenanea_leaves"), LEAVES, condition);
-		addBlock(id("helix_tree_leaves"), LEAVES, condition);
-		addBlock(id("lucernia_leaves"), LEAVES, condition);
+		BooleanSupplier leavesCondition = () -> WWBlockConfig.get().blockSounds.leafSounds;
+		addBlock(id("pythadendron_leaves"), LEAVES, leavesCondition);
+		addBlock(id("lacugrove_leaves"), LEAVES, leavesCondition);
+		addBlock(id("dragon_tree_leaves"), LEAVES, leavesCondition);
+		addBlock(id("tenanea_leaves"), LEAVES, leavesCondition);
+		addBlock(id("helix_tree_leaves"), LEAVES, leavesCondition);
+		addBlock(id("lucernia_leaves"), LEAVES, leavesCondition);
+		addBlock(id("lucernia_outer_leaves"), LEAVES, leavesCondition);
+		addBlock(id("glowing_pillar_leaves"), LEAVES, leavesCondition);
+		addBlock(id("cave_bush"), LEAVES, leavesCondition);
+		addBlock(id("end_lotus_leaf"), LEAVES, leavesCondition);
+
+		BooleanSupplier saplingCondition = () -> WWBlockConfig.get().blockSounds.saplingSounds;
+		addBlock(id("pythadendron_sapling"), SAPLING, saplingCondition);
+		addBlock(id("lacugrove_sapling"), SAPLING, saplingCondition);
+		addBlock(id("dragon_tree_sapling"), SAPLING, saplingCondition);
+		addBlock(id("tenanea_sapling"), SAPLING, saplingCondition);
+		addBlock(id("helix_tree_sapling"), SAPLING, saplingCondition);
+		addBlock(id("umbrella_tree_sapling"), SAPLING, saplingCondition);
+		addBlock(id("lucernia_sapling"), SAPLING, saplingCondition);
+
+		BooleanSupplier mushroomCondition = () -> WWBlockConfig.get().blockSounds.mushroomBlockSounds;
+		addBlock(id("mossy_glowshroom_sapling"), MUSHROOM, mushroomCondition);
+		addBlock(id("mossy_glowshroom_cap"), MUSHROOM_BLOCK, mushroomCondition);
+		addBlock(id("mossy_glowshroom_fur"), MUSHROOM_BLOCK, mushroomCondition);
+		addBlock(id("mossy_glowshroom_sapling"), MUSHROOM_BLOCK, mushroomCondition);
+		addBlock(id("small_jellyshroom"), MUSHROOM, mushroomCondition);
+		addBlock(id("jellyshroom_cap_purple"), MUSHROOM_BLOCK, mushroomCondition);
+		addBlock(id("bolux_mushroom"), MUSHROOM, mushroomCondition);
+
+		BooleanSupplier flowerCondition = () -> WWBlockConfig.get().blockSounds.flowerSounds;
+		addBlock(id("hydralux_petal_block"), FLOWER, flowerCondition);
+		addBlock(id("end_lotus_flower"), FLOWER, flowerCondition);
+		addBlock(id("tenanea_flowers"), FLOWER, flowerCondition);
 	}
 }